+++ apps/remnant2toolkit/app/(builds)/builder/_components/base-game-build-badge.tsx
@@ -3,19 +3,19 @@ import Image from 'next/image';
 
 import { Tooltip } from '@/app/_components/tooltip';
 
-export function BeginnerBuildBadge({
+export function BaseGameBuildBadge({
   unoptimized = false,
 }: {
   unoptimized?: boolean;
 }) {
   return (
-    <Tooltip content={`Denotes a gimmick build.`}>
-      <button aria-label="Badge denoting the build is a gimmick build.">
+    <Tooltip content={`Denotes a base game build.`}>
+      <button aria-label="Badge denoting the build is a base game build.">
         <Image
-          src={getImageUrl(`/badges/beginner-build-badge.png`)}
+          src={getImageUrl(`/badges/base-game-build-badge.png`)}
           width={52}
           height={60}
-          alt="Badge denoting the build is a gimmick build."
+          alt="Badge denoting the build is a base game build."
           className="h-[60px] max-h-[60px] w-[52px] max-w-[52px]"
           loading="eager"
           unoptimized={unoptimized}