+++ apps/remnant2toolkit/app/(builds)/builder/_components/build-badges.tsx
@@ -1,5 +1,6 @@
 import { useBadges } from '@/app/_hooks/use-badges';
 import { type BuildState } from '@/app/(builds)/_types/build-state';
+import { BaseGameBuildBadge } from '@/app/(builds)/builder/_components/base-game-build-badge';
 import { FeaturedBuildBadge } from '@/app/(builds)/builder/_components/featured-build-badge';
 import { GimmickBuildBadge } from '@/app/(builds)/builder/_components/gimmick-build-badge';
 import { NewBuildBadge } from '@/app/(builds)/builder/_components/new-build-badge';
@@ -14,6 +15,7 @@ interface Props {
 
 export function BuildBadges({ buildState, isScreenshotMode = false }: Props) {
   const {
+    hasBaseGameBadge,
     hasNewBadge,
     hasGimmickBadge,
     hasPopularBadge,
@@ -39,6 +41,9 @@ export function BuildBadges({ buildState, isScreenshotMode = false }: Props) {
       ) : null}
       {hasBeginnerBadge ? (
         <BeginnerBuildBadge unoptimized={isScreenshotMode} />
+      ) : null}
+      {hasBaseGameBadge ? (
+        <BaseGameBuildBadge unoptimized={isScreenshotMode} />
       ) : null}
     </>
   );